completed part 4

+++ part04/bloglist/backend/controllers/blogs.js
@@ -1,9 +1,10 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   res.json(blogs);
 });
 
@@ -16,17 +17,31 @@ blogsRouter.get("/:id", async (req, res, next) => {
   }
 });
 
+const getTokenFrom = (req) => {
+  const authorization = req.get("authorization");
+  if (authorization && authorization.startsWith("Bearer ")) {
+    return authorization.replace("Bearer ", "");
+  }
+  return null;
+};
+
 blogsRouter.post("/", async (req, res) => {
   const { title, author, url, likes } = req.body;
 
-  const user = await User.findById(body.userID);
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
     title: title,
     author: author,
     url: url,
     likes: likes,
-    user: user.id,
+    user: user._id,
   });
 
   const savedBlog = await blog.save();
@@ -37,6 +52,15 @@ blogsRouter.post("/", async (req, res) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const user = req.user;
+  const blog = Blog.findById(req.params.id);
+
+  if (blog.user.toString() !== user.id) {
+    return res
+      .status(401)
+      .json({ error: "only the user who made the blog can delete it" });
+  }
+
   await Blog.findByIdAndRemove(req.params.id);
 
   res.status(204).end();
